If applied this commit will add the program for the Restaurant Management System
Together with some updates. This program has been created primarily with the python library Tkinter to generate a graphic user interface.

This GUI allows the user to input their choice of:
1) Appetizer
2) Main Course
3) Dessert

This follows the standard 3 course meal optionality found at modern-day restaurants.

Once filled in, the program will return the cost of these items of food depending on the quantity selected, alongside the goods & services tax set to 18%, which will add up the "Total Amount" at the end.